Finalization and cleanup of the first working db implementation.

+++ modules/telestion-core/src/main/java/de/jvpichowski/rocketsound/messages/base/Pressure.java
@@ -3,10 +3,11 @@ package de.jvpichowski.rocketsound.messages.base;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.telestion.api.message.JsonMessage;
 
-public record Pressure(@JsonProperty float pressure) implements JsonMessage {
+// TODO: float not yet supported by vertx -> changed floats to double (Jan)
+public record Pressure(@JsonProperty double pressure) implements JsonMessage {
 
 	@SuppressWarnings("unused")
 	public Pressure(){
-		this(0.0f);
+		this(0.0);
 	}
 }
