+++ modules/telestion-core/src/main/java/de/jvpichowski/rocketsound/messages/sound/Spectrum.java
@@ -3,9 +3,10 @@ package de.jvpichowski.rocketsound.messages.sound;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.telestion.api.message.JsonMessage;
 
-public record Spectrum(@JsonProperty float min, @JsonProperty float max, @JsonProperty float[] data) implements JsonMessage {
+// TODO: float not yet supported by vertx -> changed floats to double (Jan)
+public record Spectrum(@JsonProperty double min, @JsonProperty double max, @JsonProperty double[] data) implements JsonMessage {
 	@SuppressWarnings("unused")
 	public Spectrum(){
-		this(0.0f, 0.0f, null);
+		this(0.0, 0.0, null);
 	}
 }
